requirements.txt
aiohttp==3.11.11
asyncio==3.4.3
bs4==0.0.2
python-telegram-bot==21.9
python-dotenv==1.0.1